feature/roomを返せるようにした

+++ grpc/domain/room_user.go
@@ -2,7 +2,7 @@ package domain
 
 type RoomUser struct {
 	Id      uint64
-	RoomId  uint
+	RoomId  uint64
 	UserUid string
 	User    User `gorm:"foreignKey"`
 }
